add prefetch links and seo enhancements
