+++ WT4Q/public/sitemap.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+  <url><loc>https://example.com/</loc></url>
+  <url><loc>https://example.com/search</loc></url>
+  <url><loc>https://example.com/login</loc></url>
+  <url><loc>https://example.com/register</loc></url>
+  <url><loc>https://example.com/weather</loc></url>
+</urlset>
